requirements.txt
anthropic
Pillow
streamlit
python-dotenv
gtts
